Extract test env data where necessary
