+++ server/src/test/java/io/spine/server/event/given/EventEnricherTestEnv.java
@@ -33,6 +33,7 @@ import io.spine.core.given.GivenUserId;
 import io.spine.people.PersonName;
 import io.spine.server.command.TestEventFactory;
 import io.spine.server.event.EventEnricher;
+import io.spine.server.event.EventEnricherTest;
 import io.spine.server.outbus.enrich.given.StringToPersonName;
 import io.spine.server.outbus.enrich.given.StringToZoneOffset;
 import io.spine.test.event.ProjectCompleted;
@@ -49,6 +50,7 @@ import org.checkerframework.checker.nullness.qual.Nullable;
 import static com.google.common.base.Preconditions.checkNotNull;
 import static io.spine.base.Identifier.newUuid;
 import static io.spine.protobuf.AnyPacker.pack;
+import static io.spine.server.command.TestEventFactory.newInstance;
 
 /**
  * @author Alexander Yevsyukov
@@ -63,6 +65,12 @@ public class EventEnricherTestEnv {
         return ProjectId.newBuilder()
                         .setId(newUuid())
                         .build();
+    }
+
+    public static Event createEvent(Message msg) {
+        final TestEventFactory eventFactory = newInstance(EventEnricherTest.class);
+        final Event event = eventFactory.createEvent(msg);
+        return event;
     }
 
     public static class GivenEventMessage {
